service/requirements.txt
# Core
numpy>=1.26,<3
pandas==2.2.2
pyarrow==17.0.0
python-dateutil==2.9.0.post0
urllib3==2.2.2
requests==2.32.3
gcsfs

# Storage / IO
duckdb==1.1.3
google-cloud-storage==2.18.2
huggingface_hub==0.24.6

# ML (training)
scikit-learn==1.5.1
lightgbm==4.5.0
joblib==1.4.2

